skeleton code improvements

+++ include/sort_kernels.cuh
@@ -1,9 +1,9 @@
 #pragma once
 #include <cuda_runtime.h>
 
-void cpu_sort(int* arr, int n);
-void merge_sort(int* arr, int n);
-void quick_sort(int* arr, int n);
-void radix_sort(int* arr, int n);
-void bitonic_sort(int* arr, int n);
-void brick_sort(int* arr, int n);
+void cpu_sort(int* arr, size_t n);
+void merge_sort(int* arr, size_t n);
+void quick_sort(int* arr, size_t n);
+void radix_sort(int* arr, size_t n);
+void bitonic_sort(int* arr, size_t n);
+void brick_sort(int* arr, size_t n);
