+++ src/bitonic_sort.cu
@@ -1,5 +1,5 @@
 #include "sort_kernels.cuh"
 
-void bitonic_sort(int* arr, int n) {
+void bitonic_sort(int* arr, size_t n) {
     return; // TODO
 }
