+++ src/cpu_sort.cu
@@ -1,6 +1,6 @@
 #include "sort_kernels.cuh"
 #include <algorithm>
 
-void cpu_sort(int* arr, int n) {
+void cpu_sort(int* arr, size_t n) {
     std::sort(arr, arr + n);
 }
