+++ src/merge_sort.cu
@@ -1,5 +1,8 @@
 #include "sort_kernels.cuh"
+#include <cuda_runtime.h>
+#include <algorithm>
+#include <iostream>
 
-void merge_sort(int* arr, int n) {
-    return; // TODO
+void merge_sort(int* arr, size_t n) {
+    return;
 }
