+++ src/radix_sort.cu
@@ -1,5 +1,5 @@
 #include "sort_kernels.cuh"
 
-void radix_sort(int* arr, int n) {
+void radix_sort(int* arr, size_t n) {
     return; // TODO
 }
